requirements.txt
-i https://pypi.org/simple
aiohttp==3.8.4 ; python_version >= '3.6'
aiosignal==1.3.1 ; python_version >= '3.7'
anyio==3.6.2 ; python_full_version >= '3.6.2'
async-timeout==4.0.2 ; python_version >= '3.6'
attrs==23.1.0 ; python_version >= '3.7'
certifi==2023.5.7 ; python_version >= '3.6'
charset-normalizer==3.1.0 ; python_full_version >= '3.7.0'
frozenlist==1.3.3 ; python_version >= '3.7'
h11==0.14.0 ; python_version >= '3.7'
httpcore==0.17.0 ; python_version >= '3.7'
httpx==0.24.0 ; python_version >= '3.7'
idna==3.4 ; python_version >= '3.5'
multidict==6.0.4 ; python_version >= '3.7'
openai==0.27.6
python-dotenv==1.0.0
python-telegram-bot==20.3
requests==2.30.0
sniffio==1.3.0 ; python_version >= '3.7'
tqdm==4.65.0 ; python_version >= '3.7'
urllib3==2.0.2 ; python_version >= '3.7'
yarl==1.9.2 ; python_version >= '3.7'
